# public/swagger-spec.json
{"openapi":"3.0.0","paths":{"/api/v1/auth/signin":{"post":{"operationId":"AuthController_signIn","summary":"Đăng nhập","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignInDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"title":"SIGN IN","type":"object","$ref":"#/components/schemas/SignInResponseDto"}}}}},"tags":["Authentication"]}},"/api/v1/account":{"post":{"operationId":"AccountController_signUpAccount","summary":"Đăng ký tài khoản","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateAccountDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"title":"SIGN UP","type":"object","$ref":"#/components/schemas/AccountDto"}}}}},"tags":["Account"]}},"/api/v1/account/me/info":{"get":{"operationId":"AccountController_getMe","summary":"Lấy thông tin account hiện tại","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"title":"GET CURRENT ACCOUNT","type":"object","$ref":"#/components/schemas/AccountDto"}}}}},"tags":["Account"],"security":[{"bearer":[]}]},"put":{"operationId":"AccountController_updateCurrentAccount","summary":"Cập nhật thông tin account hiện tại","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateAccountDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"title":"UPDATE CURRENT ACCOUNT","type":"object","$ref":"#/components/schemas/AccountDto"}}}}},"tags":["Account"],"security":[{"bearer":[]}]}},"/api/v1/file/upload/single":{"post":{"operationId":"FileController_uploadSingleFile","summary":"Upload một file","parameters":[],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"title":"UPLOAD FILE","type":"object","$ref":"#/components/schemas/UploadFileDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"title":"UPLOAD NEW FILE","type":"object","$ref":"#/components/schemas/FileDto"}}}}},"tags":["File"]}},"/api/v1/file/upload/multiple":{"post":{"operationId":"FileController_uploadMultipleFiles","summary":"Upload nhiều file","parameters":[],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"files":{"type":"array","items":{"type":"string","format":"binary"}}}}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"title":"UPLOAD NEW FILES","type":"array","items":{"$ref":"#/components/schemas/FileDto"}}}}}},"tags":["File"]}},"/api/v1/file/{id}":{"get":{"operationId":"FileController_getFileById","summary":"Lấy thông tin một file","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"title":"GET A FILE","type":"object","$ref":"#/components/schemas/FileDto"}}}}},"tags":["File"],"security":[{"bearer":[]}]},"put":{"operationId":"FileController_updateFileById","summary":"Cập nhật thông tin một file","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateFileDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"title":"UPDATE A FILE","type":"object","$ref":"#/components/schemas/ApiSuccess"}}}}},"tags":["File"],"security":[{"bearer":[]}]},"delete":{"operationId":"FileController_deleteFileById","summary":"Xóa một file","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"title":"DELETE A FILE","type":"object","$ref":"#/components/schemas/ApiSuccess"}}}}},"tags":["File"],"security":[{"bearer":[]}]}},"/api/v1/country":{"get":{"operationId":"CountryController_getAllCountries","summary":"Lấy danh sách country","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"title":"Get All Countries","type":"array","items":{"$ref":"#/components/schemas/CountryDto"}}}}}},"tags":["Country"]}}},"info":{"title":"Vietnamese Learning","description":"API for Vietnamese Learning","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"SignInDto":{"type":"object","properties":{"username":{"type":"string","description":"Must input username"},"password":{"type":"string","description":"Must input password"}},"required":["username","password"]},"SignInResponseDto":{"type":"object","properties":{"accessToken":{"type":"string","description":"This is access token"}},"required":["accessToken"]},"FileDto":{"type":"object","properties":{"id":{"type":"string","description":"This is file's id"},"fieldName":{"type":"string","description":"This is file's field name"},"originalName":{"type":"string","description":"This is file's field name"},"encoding":{"type":"string","description":"This is file's encoding"},"mimeType":{"type":"string","description":"This is file's mimetype"},"fileName":{"type":"string","description":"This is file's file name"},"url":{"type":"string","description":"This is file's url"},"size":{"type":"number","description":"This is file's size"},"objectId":{"type":"string","description":"This is object's id"},"objectType":{"type":"string","description":"This is object's type"}},"required":["id","fieldName","originalName","encoding","mimeType","fileName","url","size","objectId","objectType"]},"CountryDto":{"type":"object","properties":{"id":{"type":"string","description":"This is id"},"name":{"type":"string","description":"This is name"},"nameEN":{"type":"string","description":"This is English name"}},"required":["id","name","nameEN"]},"AccountDto":{"type":"object","properties":{"fullName":{"type":"string","description":"This is full name"},"phone":{"type":"string","description":"This is phone"},"email":{"type":"string","description":"This is email"},"avatar":{"description":"This is avatar","allOf":[{"$ref":"#/components/schemas/FileDto"}]},"dob":{"type":"string","description":"This is dob"},"gender":{"type":"string","description":"This is gender"},"nationality":{"description":"This is nationality","allOf":[{"$ref":"#/components/schemas/CountryDto"}]},"username":{"type":"string","description":"This is username"},"password":{"type":"string","description":"This is password"}},"required":["fullName","phone","email","avatar","dob","gender","nationality","username","password"]},"UpdateAccountDto":{"type":"object","properties":{"fullName":{"type":"string","description":"Input full name"},"phone":{"type":"string","description":"Input phone"},"avatar":{"type":"string","description":"Input avatar"},"dob":{"type":"string","description":"Input dob"},"gender":{"type":"string","description":"Input gender"},"nationality":{"type":"string","description":"Input nationality"},"password":{"type":"string","description":"Input password"}},"required":["fullName","phone","avatar","dob","gender","nationality","password"]},"CreateAccountDto":{"type":"object","properties":{"fullName":{"type":"string","description":"Input full name"},"phone":{"type":"string","description":"Input phone"},"email":{"type":"string","description":"Input email"},"avatar":{"type":"string","description":"Input avatar"},"dob":{"type":"string","description":"Input dob"},"gender":{"type":"string","description":"Input gender"},"nationality":{"type":"string","description":"Input nationality"},"password":{"type":"string","description":"Input password"}},"required":["fullName","phone","email","avatar","dob","gender","nationality","password"]},"UploadFileDto":{"type":"object","properties":{"file":{"type":"string","description":"Must input file","format":"binary"}},"required":["file"]},"UpdateFileDto":{"type":"object","properties":{"objectId":{"type":"string","description":"This is object's id"},"objectType":{"type":"string","description":"This is object's type"}},"required":["objectId","objectType"]},"ApiSuccess":{"type":"object","properties":{"statusCode":{"type":"number"},"message":{"type":"string"}},"required":["statusCode","message"]}}}}